Rearrange into cajita, core, and utils
