+++ cajita/unit_test/CMakeLists.txt
@@ -52,7 +52,7 @@ macro(Cajita_add_tests)
       list(APPEND CAJITA_UNIT_TEST_NUMTHREADS ${_nt})
     endif()
   endforeach()
-  set(CAJITA_UNIT_TEST_MAIN ../../core/unit_test/mpi_unit_test_main.cpp)
+  set(CAJITA_UNIT_TEST_MAIN ../../utils/unit_test/mpi_unit_test_main.cpp)
   foreach(_device ${CAJITA_TEST_DEVICES})
     set(_dir ${CMAKE_CURRENT_BINARY_DIR}/${_device})
     file(MAKE_DIRECTORY ${_dir})
@@ -62,7 +62,7 @@ macro(Cajita_add_tests)
       file(APPEND ${_file} "#include <tst${_test}.hpp>\n")
       set(_target Cajita_${_test}_test_${_device})
       add_executable(${_target} ${_file} ${CAJITA_UNIT_TEST_MAIN})
-      target_include_directories(${_target} PRIVATE ${_dir} ${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_CURRENT_SOURCE_DIR}/../../core/unit_test)
+      target_include_directories(${_target} PRIVATE ${_dir} ${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_CURRENT_SOURCE_DIR}/../../utils/unit_test)
       target_link_libraries(${_target} PRIVATE Cajita cabana_cajita_gtest)
       if(CAJITA_UNIT_TEST_MPI)
         foreach(_np ${CAJITA_UNIT_TEST_MPIEXEC_NUMPROCS})
