+++ cmake/CabanaConfig.cmakein
@@ -30,7 +30,8 @@ if(Cabana_ENABLE_ARBORX)
   find_dependency(ArborX REQUIRED)
 endif()
 set(Cabana_ENABLE_CAJITA @Cabana_ENABLE_CAJITA@)
+include("${CMAKE_CURRENT_LIST_DIR}/CabanaUtilsTargets.cmake")
 if(Cabana_ENABLE_CAJITA)
   include("${CMAKE_CURRENT_LIST_DIR}/CajitaTargets.cmake")
 endif()
-include("${CMAKE_CURRENT_LIST_DIR}/Cabana_Targets.cmake")
+include("${CMAKE_CURRENT_LIST_DIR}/CabanaCoreTargets.cmake")
